+++ src-tauri/src/utils/app/mod.rs
@@ -3,6 +3,8 @@ pub mod builder;
 pub mod notification;
 pub mod platform;
 pub mod ui;
+
+#[cfg(feature = "updater")]
 pub mod updater;
 
 #[cfg(not(feature = "web-server"))]