+++ src-tauri/src/utils/rclone/mod.rs
@@ -4,5 +4,7 @@ mod extractor;
 pub mod mount;
 pub mod process_common;
 pub mod provision;
-pub mod updater;
 pub mod util;
+
+#[cfg(feature = "updater")]
+pub mod updater;