+++ src-tauri/src/utils/types/mod.rs
@@ -8,4 +8,6 @@ pub mod origin;
 pub mod rclone;
 pub mod remotes;
 pub mod state;
+
+#[cfg(feature = "updater")]
 pub mod updater;